Additions and subtractions to the publish script.
