+++ manifest.json
@@ -20,7 +20,7 @@
                 "password_var": "ACRE_STEAM_PASSWORD",
                 "project_id": 522695873,
                 "release_dir": "P:\\idi\\acre\\release\\@acre2",
-                "compat": []
+                "release_text": "P:\\idi\\acre\\release\\@acre2\\README.md"
             }
         ]
     }
